pyproject.toml
[project]
name = "agent2"
version = "0.1.0"
description = "The production runtime for AI agents. Schema in, API out."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
dependencies = [
    "pydantic-ai>=0.1",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.0",
    "httpx>=0.27",
    "redis>=5.0",
    "pyyaml>=6.0",
    "langfuse>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-httpx>=0.30",
    "ruff>=0.5",
]
# Promptfoo is a Node.js CLI tool (npm install -g promptfoo).
# This group contains Python helpers used by custom eval scripts.
promptfoo = [
    "httpx>=0.27",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["shared"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests that require Docker services",
]

[tool.ruff]
line-length = 120
target-version = "py312"
